Fix some invalid reads and writes in shortpath.
This changes the behavior of shortpath: a newly allocated memory block will be returned. It has to be deallocated with free.

+++ src/libmhe/shortpath.h
@@ -13,6 +13,6 @@
 #ifndef _shortpath_h_included
 #define _shortpath_h_included
 
-char *shortpath(const char *path, unsigned maxlen, char *home);
+char *shortpath(const char *path, size_t maxlen, char *home);
 
 #endif
